requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.39.0
datasets>=2.18.0
sentence-transformers>=2.6.0
accelerate>=0.28.0
pyyaml>=6.0.1
pillow>=10.2.0
tqdm>=4.66.0
numpy>=1.26.0
jupyter>=1.0.0
